+++ tests/test_addic7ed.py
@@ -272,6 +272,22 @@ def test_query_parsing(episodes):
 
 @pytest.mark.integration
 @vcr.use_cassette
+def test_query_parsing_quote_dots_mixed_case(episodes):
+    video = episodes['marvels_agents_of_shield_s02e06']
+    with Addic7edProvider() as provider:
+        subtitles = provider.query(video.series, video.season)
+    subtitle = [s for s in subtitles if s.download_link == 'updated/10/93279/9'][0]
+    assert subtitle.language == Language('por', country='BR')
+    assert subtitle.hearing_impaired is False
+    assert subtitle.page_link == 'http://www.addic7ed.com/serie/Marvel%27s_Agents_of_S.H.I.E.L.D./2/6/A_Fractured_House'
+    assert subtitle.series == video.series
+    assert subtitle.season == video.season
+    assert subtitle.episode == video.episode
+    assert subtitle.version == 'KILLERS'
+
+
+@pytest.mark.integration
+@vcr.use_cassette
 def test_query_year(episodes):
     video = episodes['dallas_2012_s01e03']
     with Addic7edProvider() as provider: