+++ HISTORY.rst
@@ -8,6 +8,7 @@ Changelog
 * Fix library usage example in README
 * Support python 3.5
 * Improve matching on titles
+* Fix for series name with special characters in addic7ed provider
 
 
 1.0.1